pyproject.toml
[tool.poetry]
name = "logflow"
version = "0.1.0"
description = "A flexible ETL system for log processing"
authors = ["OpenHands"]
readme = "README.md"
packages = [{include = "logflow"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pyyaml = "^6.0"
structlog = "^23.1.0"
click = "^8.1.3"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
pydantic = "^1.10.7"
aiofiles = "^23.1.0"
aiokafka = "^0.8.0"
elasticsearch = "^8.7.0"
opensearch-py = "^2.0.0"
aiobotocore = "^2.5.0"
python-dotenv = "^1.0.0"
prometheus-client = "^0.16.0"
jinja2 = "^3.1.2"
user-agents = "^2.2.0"
geoip2 = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"
sphinx = "^6.1.3"
sphinx-rtd-theme = "^1.2.0"

[tool.poetry.scripts]
logflow = "logflow.cli.commands:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
